JavaSE/SDA/wzorceProjektowe/src/test/java/pl/anicos/patterns/exercises/abstractfactory/AbstractFactoryApplicationTest.java
package pl.anicos.patterns.exercises.abstractfactory;

import org.junit.Assert;
import org.junit.Test;
import pl.anicos.patterns.exercises.abstractfactory.buttons.MacOSButton;
import pl.anicos.patterns.exercises.abstractfactory.buttons.WindowsButton;
import pl.anicos.patterns.exercises.abstractfactory.checkboxes.LinuxCheckbox;
import pl.anicos.patterns.exercises.abstractfactory.checkboxes.MacOSCheckbox;
import pl.anicos.patterns.exercises.abstractfactory.checkboxes.WindowsCheckbox;
import pl.anicos.patterns.exercises.abstractfactory.buttons.LinuxButton;
import pl.anicos.patterns.exercises.abstractfactory.factories.LinuxFactory;
import pl.anicos.patterns.exercises.abstractfactory.factories.MacOSFactory;
import pl.anicos.patterns.exercises.abstractfactory.factories.WindowsFactory;

public class AbstractFactoryApplicationTest {

    @Test
    public void shouldPaintMacOsUi() {
        //given
        AbstractFactoryApplication testObj = new AbstractFactoryApplication(new MacOSFactory());
        //when
        String[] paint = testObj.paint();
        //then
        Assert.assertEquals(MacOSButton.YOU_HAVE_CREATED_MAC_OSBUTTON, paint[0]);
        Assert.assertEquals(MacOSCheckbox.YOU_HAVE_CREATED_MAC_OSCHECKBOX, paint[1]);
    }

    @Test
    public void shouldPaintWindowsUi() {
        //given
        AbstractFactoryApplication testObj = new AbstractFactoryApplication(new WindowsFactory());
        //when
        String[] paint = testObj.paint();
        //then
        Assert.assertEquals(WindowsButton.YOU_HAVE_CREATED_WINDOWS_BUTTON, paint[0]);
        Assert.assertEquals(WindowsCheckbox.YOU_HAVE_CREATED_WINDOWS_CHECKBOX, paint[1]);
    }

    @Test
    public void shouldPaintLinuxUi() {
        //given
        AbstractFactoryApplication testObj = new AbstractFactoryApplication(new LinuxFactory());
        //when
        String[] paint = testObj.paint();
        //then
        Assert.assertEquals(LinuxButton.YOU_HAVE_CREATED_LINUX_BUTTON, paint[0]);
        Assert.assertEquals(LinuxCheckbox.YOU_HAVE_CREATED_LINUX_CHECKBOX, paint[1]);
    }

}